requirements.txt
asgiref==3.6.0
black==23.3.0
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
coreapi==2.3.3
coreschema==0.0.4
Django==4.2.1
django-debug-toolbar==4.0.0
django-filter==23.2
djangorestframework==3.14.0
drf-yasg==1.21.5
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
Markdown==3.4.3
MarkupSafe==2.1.2
mypy-extensions==1.0.0
packaging==23.1
pathspec==0.11.1
platformdirs==3.5.0
pytz==2023.3
requests==2.30.0
ruamel.yaml==0.17.25
ruamel.yaml.clib==0.2.7
sqlparse==0.4.4
uritemplate==4.1.1
urllib3==2.0.2
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.2.2
PyJWT==2.8.0
jwt~=1.3.1